HPCC-25484 Move daliadmin helper functions for ESP WsDali
Rename daliadmin.cpp to daadmin.cpp. Create a new daliadmin.cpp.
Move the main(), usage(), and related code from daadmin.cpp back
to new daliadmin.cpp. Add the code to create a new library
daliadminlib which contains the helper functions in daadmin.cpp.

This is the 1st commit. This commit does not have new daliadmin.cpp
which will be in the 2nd commit.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>