HPCC-25708 Auto populate file level @compressedSize based on parts
@size was auto-calculated and populated based on the part meta info
when creating a distributed file, but @compressedSize wasn't.
Instead, the engines had code to manually calculate and populate
this (and in some cases @size too). And in a couple of cases they
did not populate @compressedSize, so it was missing.

Fix by always calculating and populating in DFS and removing
the engine code that did it.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>