HPCC-25787 WUQuery very slow if workunit(s) have lots of app values
The implementation of cached application values was pathologically
slow when large numbers existed (e.g. because a job had sprayed
thousands of files).

The implementation held the parent tree and build an index xpath
to each element ("*[<n>]").
This cause a linear search for each item, for every access method
of each IConstWUAppValue, i.e. each queryName() and queryValue()
call.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>