HPCC-26148 sigbus error on cache warming code
It's not clear why the error happened, though research suggests it may happen
if an NFS mount disconnects.

Recoded so that the warming is done in a separate process, and SIGBUF errors
are caught and reported.

Do not bother touching mmap for in-cache index pages, which will be loaded
anyway by subsequent code inside Roxie process.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>