HPCC-26477 Change schemaLocation value and default behavior
Change the default behavior to omit returning the schemaLocation for normal
method responses (HTTP OK). It is returned unconditionally in Soap Fault
responses.

Correctly encode the value when it is returned. Change the returned value to
conform to the standard, which is a list of URI pairs of namespace and schema
location. The hostname of the location URI is pulled from the
wsdlServiceAddress attribute of the service binding, and if not present the
schemaLocation is omitted.

Add a DESDL binding configuration attribute at
Definition[@returnSchemaLocationOnOK] to enable schemaLocation return in normal
method responses.

Use member initialization to set default schemaLocation return state.

Signed-off-by: Terrence Asselin <terrence.asselin@lexisnexisrisk.com>