HPCC-25964 Fix disk read/write stats
Disk read and write stats. were wildly incorrect, due
to a bug in the way they were serialized.
Each time a worker collected disk stats, it was rolling them
up on to the previously collected stats.
Fix by ensuring that the total read/written is set (not merged)
into the statistic collection that is serialized back to the
manager.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>