HPCC-26810 Fix use of unused PR docker image tag.
The build-and-push step used to test the containerized
build of the PR attributed a tag based on the github repository
name.
If the repository name had upper case characters, the docker
build process would complain about an invalid tag.
This tag was not actally used (image not pushed), so the fix
was to remove the tag from the step.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>