HPCC-26842 Fix ptree external prop. serialization issue
An external (disk based) property from Dali was not being
serialized to the client if it had not been read before.
NB: this is only an issue when serializing ptree's to clients
via non client based SDS routes (i.e. not via ISDSConnection's).
e.g. via getFileTree in SDS for foreign files.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>