HPCC-26884 Initialize ESP session information when binding is loaded
The exiting code initializes ESP session information when ESP is started. That
may not work for dynamically loaded ESP bindings. In this fix:
1. Move the code for initializing the CSessionCleaner out of the code for the
session information initialization. Initialize CSessionCleaner when ESP is
started.
2. When a binding is added, if dali is available. initialize session
information in dali; if dali is detached, initialize session
information when dali is attached.

Revise based on review:
1. Add thread safety for binding port list
2. Rename getSDSSessionTree
3. Fix thread safety for binding port list

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>