HPCC-26680 ESDL tool not honoring exceptions_inline
Generated wsdl and xsd from the command-line esdl tool don't include
the inline Exceptions element in response elements that are marked in
the ecm files with the exceptions_inline attribute.

Update the xslt file that creates the xsd/wsdl to output the exceptions
element inline by default when it is in the ecm markup. Provide a parameter
to turn off exceptions inline in the stylesheet and in the esdl tool. For
compatibility with existing DESDL service behavior, modify the ESDL binding
code to turn off exceptions inline until full support is enabled with ticket
(HPCC-27342).

Correct the generated wsdl to define the wsdl namespace.
Add esdl test cases to exercise exceptions_inline behavior.

Signed-off-by: Terrence Asselin <terrence.asselin@lexisnexisrisk.com>