HPCC-27384 Roxie may core if processing two reloads at same time
Deploying a query typically causes Roxie to reload twice (once from the
notification of the dali queryset, once called explicitly). This may core if a
query is suspended because of a missing file, as the dummy queryFactory that
is created ends up being updated on both threads at the same time.

Rework so that the updating of the error message is done only on
construction/load, and the field is immutable thereafter.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>