HPCC-27170 Fix remote index filter/limit issues
Fix a number of issues related to reading indexes remotely
(via dafilesrv streaming), where filtering or limits are present.

The indexread/count/normalize activities optionally pass a threshold
limit to the remote side, to suppress excessive unecessary rows
being serialized back.
This is total remoteLimit disabled if there are is any post filtering,
because it is not known how many remote rows will eventually be
needed.
Three issues fixed:
1) IndexCount was using helper->hasMatchFilter() but hasFilter()
was set instead. This was causing it to send a remoteLimit, and
therefore not receive rows back which would have matched the post-
filter.
2) IndexNormalize was also basing remoteLimit on hasMatchFilter().
However, it neither sets hasFilter or hasMatchFilter. Change to
assume no remoteLimit in indexnormalize case.
3) getRowLimit was being used as a remoteLimit boundary, meaning
it would falsely hit the limit, when post filters would reduce
it below that limit.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>