HPCC-27621 LDAP regression connecting to secure AD
When connecting to LDAP AD query server parameters, the code used to
attempt anonymous bind on nonsecure (389) port if the secure port failed.
This code was erroneously removed (https://track.hpccsystems.com/browse/HPCC-27522),
as it has now been determined it needed to query the domain name other to
allow the component to properly initialize.

Signed-off-by: Russ Whitehead <russwhitehead@yahoo.com>