HPCC-27597 Fix JSON output file with nested records has extra newlines
The current CommonJsonWriter output is valid JSON, but scalars can end
up with ugly formatting after a nested record.  The real problem is that
this confuses despray and we end up losing data.  I think despray relies
on specific newline placement not just valid JSON.

A proper fix for the nestChildren feature (which is a flag parameter of
outputBeginNested()) would be complicated.

So would a fix for despray which currently relies on how the newlines
are placed.

The fix in this PR is meant to be a quick, simple, low risk fix for the
ECL output scenario only so that despray works as expected.

Signed-off-by: Anthony Fishbeck <anthony.fishbeck@lexisnexis.com>