HPCC-27651 Compress external string values
Convert string blobs to compressed format, and denote with
extra flag, so that they can be converted back to uncompressed
on read.
This preserves a common format although if Dali was rolled
back, anything accessing these values via the IPT::queryProp
interface, would not see the correct data (i.e. compressed raw
value)

Also fix a minor bug where a property remained flagged as binary,
if it was originally set as binary then set as non-binary,
which had the side effect of returning an extra terminator byte
in the property value.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>