HPCC-27687 Check permission before accessing Others Workunits
1. When a WU is specified in WsWorkunits.WUQuery, the user
permission should be verified before accessing Others Workunits.
2. In the WsSMC service and WsWorkunits service, a new flag
CanAccess is added when the services report a list of ECL
workunits. The flag is set to false for Others Workunits if the
ESP client does not have the permission to access Others
Workunits. ECLWatch may use the flag to decide whether a weblink
for workunit details should be added or not.

Revise based on review

1. Change CanAccess to NoAccess.
2. Remove the version checks since they will be done by the
generated serialization code.
3. Change the ambiguous exception.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>