HPCC-27695 Fix ESP leaks and error reporting
Generated ESP CClient method implementations leak when there are exceptions
thrown in their call to post(). Fix by reference counting the response
object.

SSL plugin is loaded each time the HttpClientContext is created, but not
managed. Instead, load once into an Owned pointer in the module and ensure
it is destroyed on module exit.

Update error messages to have useful information when SOAP client doesn't get
a response from the server.

These issues are seen in the context of the esdl tool calling to the
WsEsdlConfig service.

Signed-off-by: Terrence Asselin <terrence.asselin@lexisnexisrisk.com>