HPCC-27714 LDAPEnv generated admin usernames too long
LDAP restricts usernames to 20 characters or less. Because the environment name
is part of the generated username, these names can easily be too long. Therefore
first try to create the username as is, then try truncating to 20 characters and
retrying.  This can fail if the same username exists elsewhere (where we cannot
search), so generate a numeric suffix (1-99) and add that to the generated username
So, we try HPCC_LongEnvironmentName, HPCC_LongEnvironment, HPCC_LongEnvironmen1,
HPCC_LongEnvironmen2, and so forth

Signed-off-by: Russ Whitehead <russwhitehead@yahoo.com>