HPCC-27374 Do not throw the exception when deleting Unknown state WU
In the existing code, the exception ("Please abort before deleting
this workunit") will be thrown when deleting a WU with Unknown state.
When deleting a WU, all the data about the WU will be deleted,
including the state attribute. Since the state attribute will be
deleted, there is no need to change the state before deleting it.
In this fix, the code is changed to not throw the exception.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>