HPCC-27739 Fix remote plane mapping issue effecting numDevices
When mapping from remote to local planes, it must also rename
the Cluster/@name, which is used by code that looks up the
numDevices.

Without this fix, it will fail to find the correct local plane
(or use the wrong local plane, that happens to match the remote
plane name), and get the numDevices wrong.
As a consequence, it will result in the wrong local paths being
constructed and cause 'Could not open logical file ... in any of
these locations: ...' errors

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>