HPCC-27866 Fix duplicated items in WsWorkunits.WULogFileRequest
ESPStruct WUFileOption was added to the WULogFileRequest by
ps://github.com/hpcc-systems/HPCC-Platform/pull/15992.
It should not be added there because several items (IPAddress,
LogDate, etc.) inside the WUFileOption are already in the
WULogFileRequest. A user may not know those items inside
WUFileOption are not used for the WsWorkunits.WULogFileRequest.

Revise based on review: modified the readWUComponentLogOptionsReq().

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>