HPCC-27993 Fix directio regression and striping support
1) The changes in HPCC-25041 (in 8.8.0) meant the dafilesrv
directio component failed because it's configuration option
was not picked up due to a casing typo in the dafilesrv
type logic.
2) Clients didn't handle striping correctly. Fix by
ensuring the serialize plane is available and the time of
directory manipulation.
3) Dafilesrv unnecessarily rejecting streaming request in a 'directio'
application setup. In a K8s directio application, it should support
all file operations (std i/o and streaming requests).

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>