HPCC-28036 Fix k8s despray issues from plane with hosts
Trying to despray from a plane with hosts (e.g. for despraying
to baremetal) failed because the service code expected to see
a @hostsGroup attribute in the plane.
Instead if should look for 'hosts', which are generated
automatically at initialization time for any planes that have
@hostGroup definitions.

This PR:
1) changes service code to universally look for "hosts" not
@hostGroup
2) Ensures all client programs call initializeStorageGroups()
which ensures that planes populate "hosts" where needed.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>