HPCC-28136 Permit disabled: false under sasha services
Prior to this change, setting disabled to false meant
the conditional logic ignored the setting, but the range
saw the "disabled" as a key in sasha services dictionary,
and as an unknown sasha service type.
This PR refactors the use of sasha services to remove
all (if disabled: true) or strip the pseudo 'disabled'
service name if false.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>