HPCC-28145 Fix path check issue in FileSpray.FileList
In https://github.com/hpcc-systems/HPCC-Platform/pull/15181,
a path verification function was added to check whether a
client is allowed to access pre-configured file pathes. To
pass the path verification. a trailing slash is needed inside
the file path. That breaks the existing FileSpray.FileList when
the file path has no trailing slash and the OS type is not
specified in the FileSpray.FileList request. The existing code
depends on the OS type to check/append the trailing slash before
the path verification. In this PR, the code is changed to check/
append the trailing slash based on the given file path.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>