HPCC-26831 Support vault secret access from ECL in bare metal
After this change, to configure vaults for a baremetal enviroment add
vault information such as:

<vaults>
  <ecl name="my-ecl-vault" url="http://localhost:8200/v1/secret/data/ecl/${secret}" kind="kv-v2" client-secret="vaultSecret"></ecl>
  <eclUser name="my-eclUser-vault" url="http://localhost:8200/v1/secret/data/eclUser/${secret}" kind="kv-v2" client-secret="vaultSecret"></eclUser>
</vaults>

To the environment.xml file at the path: /Environment/Software/vaults

Notice in the vault config xml, the mention of "client-secret".
That specifies where the token to access vault itself is stored.
In the above example client-secret="vaultSecret" so the hpcc will look
for the vault token in:

/opt/HPCCSystems/secrets/system/vaultSecret/token

With the token in that file, make sure there is no whitespace, newlines, etc.

Signed-off-by: Anthony Fishbeck <anthony.fishbeck@lexisnexis.com>