HPCC-28164 Report Queues and Roxie Targets in ws_resources
1. Add new method TargetQuery to the ESP WsResources service. The
method is used to report the information about HPCC Queues and the
names of the roxies which do not listen to port zero.
2. CTpWrapper::getServices() is revised to read the roxie services
and esp services in bare metal environment. With this change, the
information about those services is reported in WsResources.ServiceQuery
method.
3. The ESP deployment script for bare metal environment is revised to
load the WsResources into bare metal ESP.

Revise based on review:
1. Add 2 point checks.
2. Add the canDeployQueries() and related code.
3. Revise the isQueriesOnly() and related code.
4. Report TLS Secure for bare metal environment.

Revise based on review:
Rename the RTTQueries to RTTPublished;
Rename the RTTQueue to RTTQueued;
Rename the isQueriesOnly() to onlyPublishedQueries();
Rename the canDeployQueries() to canPublishQueries();
Fixed 2 readRoxieTargetType() calls.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>