HPCC-28007 Allow dafilesrv limits/throttling to be configured in K8s
And increase default pool and parallel request limits.

Also:
+ Change spray-service so using cluster visibility (was local).
The service is only needed by the in-cluster dfuserver.
+ Add retry loop (with logging) when spray service is used.
This is to catch the case when the service is overwhelmed dealing
with a lot of concurrent requests already (as configured).
+ Allow spray service to use TLS.
Previously dafilesrv was hardcoded so only the "stream" service
used TLS in containerized mode.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>