HPCC-28185 Fix external sink serialization
A slave external sink activity was spuriously serializing an
extra edge output (always equal to 0). This caused the master
to interpret this as the start of trailing new activity with
an id of 0, which doesn't exist and causes a warning log message
to be emitted.
Cause: the implementation was incorrectly adding an output edge
instance in it's ctor.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>