HPCC-28341 Fix EspBinding to allow no dali with no AuthDomains
Do not rely on readAuthDomainCfg() to determine the AuthType for checking
when no dali is active. Because it defaults to AuthTypeMixed when no
AuthDomains are configured it must be called after handling bindings with
no dali active. However, maintain the assumption that when authType value
is missing that it defaults to AuthTypeMixed.

The configmgr is one ESP without AuthDomains configured that cannot be
assumed to be using AuthTypeMixed.

Signed-off-by: Terrence Asselin <terrence.asselin@lexisnexisrisk.com>