HPCC-28110 Fix issues reading/copying multi cluster file
Reading a foreign file that exists on multiple clusters, caused
problems at the client side.
The deserialized groups referencing the foreign locations, are
instatiated in the local environment as 'anon' groups.
Then, when the deserialized IFileDescriptor info is used
to create a IDistributedFile and later used again to
serialize, the 'anon' groups are seen as a local match and
used in place of the original groups, and as a side-effect
also replace the remote base directories with the local env.
default directories, appending 'unknown' to the end of them.

This in turn causes the adjustClusterDir/replaceClusterDir
functions, to misbehave and fail to substitue in the preferred
cluster.
The consequence is that the prefered cluster is used, but with
the wrong base directory, and the physical file parts are not
found.

Fix by ensuring the original deserialized foreign group is
preserved in the IClusterInfo's.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>