HPCC-28407 Add missing forcePerms and mount secrets
eclagent and thor agent did not have an init container
to force perms where required, or mount secrets.
One consequence was that remote dfs file lookups failed
could fail on systems where the temp storage needed
forcePermissions to work, because of the way secrets
for the dfs service are fetched.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>