HPCC-28410 Fix the code for filtering out duplicated LDAP BaseDNs
The BaseDNs are stored in the ws_access as a member variable. The
BaseDNs are read from settings into the variable when the variable
is used at the first time. There is the code to filter out some
duplicated items in the setting. The code may not work when multiple
threads access the code. In this PR, a CriticalSection is added so
that only one thread can access the code at the same time.

Revise based on review:
Add the code to call setBasedns() in permissionAddInputOnAccount().

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>