HPCC-28430 Not set port number in redirect URL when not available
When ESP receives a request with an invalid session ID, ESP will
redirect the request to the Session Login page using the URL from
the HTTP Host header in the request. If the URL does not contain
a port number, ESP should not add a port number.

Revise based on review: clean the ; and virtual from the line 153-155
in the httptransport.ipp.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>