HPCC-28942 Check dropzone scope access in ESP FileSpray
1. Add getDropZoneScopePermissions() to CDistributedFileDirectory.
2. Add getDropZoneScopePermissions() to the ESP SMC lib.
3. Check dropzone scope access using the getDropZoneScopePermissions().

Revise bassed on review
1. Change the variable names from isExternalPlane to isPlane.
2. Add more information to the error messages.
3. Check isDir before checking scope access for the
files in subfolders.
4. Remove a comment.
5. Remove the extra newline.
6. When searching dropzone files, do my own recursion into subdirs
and check the scope permissions.
7. When deleting given dropzone files, add the code to check the
scope permissions for every given files.
8. Rename one isExternalPlane to isPlane.
9. Remove one unusued param in getDropZoneScopePermissions() and change
   the related code for calling the getDropZoneScopePermissions().
10. Remove the extra fname length checks.
11. Remove the unused pathSep.
12. Pass in nullptr for the relDir.
13. Change the code to use getSecAccessFlagName().
14. Clean the checkDropZoneFileScopeAccess().

More changes based on review
1. Unify the getDropZoneScopePermissions() in the dali
base lib to only accept dropzone name.
2. Add getPlaneHosts() to the dali base lib.
3. Remove the makeDropZoneScope() and getDropZoneInfoByIP().
4. Unify the onDespray() using the findDropZonePlane().
5. Add findDropZonePlaneName() to find out dropzone plane
name from dropzone host and path using findDropZonePlane().
6. Modify the getDropZoneScopePermissions() in the SMC lib
to support the getDropZoneScopePermissions() in the dali base lib.
7. Modify the getDropZoneScopePermissions() calls in:
CFileSpraySoapBindingEx::downloadFile
CFileSpraySoapBindingEx::onStartUpload
CFileSprayEx::onFileList
CFileSprayEx::onDropZoneFileSearch
CFileSprayEx::onDropZoneFiles
CFileSprayEx::onDeleteDropZoneFiles
8. Modify onFileList() to allow the host not been specified
(retrieve the files from all of the hosts in a dropzone).
9. Modify the getDropZoneFiles() so that it can be used by
multiple functions.
10. In thee searchDropZoneFiles(), do not use the nameFilter when
calling the directoryFiles().
11. Revise the getPlaneHostGroup().
12. Remove the misleading 'true' when calling getScopes().
13. Change the code to call the splitDirTail().
14. Fix a bug: when a dropzone host is not requested in the
FileSpray.FileList (the method will report the files in
all hosts in a dropzone), the validateDropZonePath() should
be called for all hosts. The code should not check the log
access.
15. Use localhost in FileList when no hostGroup and hosts settings.
16. Revise getPlaneHostGroup() based on review.
17. Add user ID and allowed permission to error messages based on review.

Fix regression for listing log files using FileList:
Before the dropzone scope check is added to the ESP FileSpray.FileList,
the FileList can also be used to retrieve a list of log files. In this
PR, the function of retrieving a list of log files is added back to the
FileList.
1. Rename the addDropZoneFile() to addPhysicalFile(). Rename/modify the
getDropZoneFiles() to getPhysicalFiles(). Both functions can be used
for both dropzone files and log files.
2. If a user does not specify a dropzone name and the specified path is
a valid log path, the FileList calls the getPhysicalFiles() to retrieve
a list of log files.

The getDropZoneScopePermissions should check all scopes:
Prior to this change, it called getScopes which returns only
the leading scopes.

Ensure isDropZoneRestrictionEnabled works as before.

Use relative dropzone path to get dropzone scope permissions.

More changes based on review:
1. Log an MCoperatorWarning when a host/IP being used to spray/
despray/list file in a cloud system.
2. Combine the 2 if(dstxml.length() == 0) in the Despray().
3. Change the PROGLOG.
4. Change the code related to list logs.
5. Several minor changes.

More changes based on review:
1. Add the mustMatch flag to the findDropZonePlane. To match
with another PR, also add the code to check if the path is empty.
2. Remove the findDropZonePlaneName().
3. Revise readAndCheckSpraySourceReq() to check srcIP and clean the code.
4. Use the isContainerized() to replace the ifdef _CONTAINERIZED.
5. Move the validateConfigurationDirectory(). Add the code to
throw exception. Add the comment about the dual purpose of the FileList.
6. Fix two typos.
7. Replace throwOrLogDropZonePlaneNotFound90 with getAndValidateDropZone()

More changes based on review:
1. Add the host/path to the 'DropZone not found' exception.
2. Change 1 comment.
3. Remove the MCoperatorWarning log calls when a plane is
resolved by hostname in containerized. And the comments
about 'should not resolve plane by hostname in containerized'.
4. Replace the 2 'if (di->isDir())' to 1
5. Pass null path to findDropZonePlane when relative path is used for
despray.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>