HPCC-28209 Fix Process Down warning for thor slaves in Target Cluster Preflight page
This PR fixes 2 issues in the preflight function for a bare-metal
environment. 1. incorrect pid file names are used in the existing
preflight function to check Thor slave processes and Thor spare
processes. In this fix, the code is changed to use the node sequence
numbers in the nodeGroup to create the pid file names for those processes.
2. The existing code logs a 'Cannot find node group ...' message when
it cannot find the node group for a Thor spare process, even if the
Thor cluster does not have a spare process. This issue is fixed.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>