HPCC-28601 AddFileScope on AAD reports User Account Already Exists
Because AAD adds the current user to a newly created FileScope OU, it is
not an error that that user's ACE entries exist in the Access Control List (ACL).
This PR adds some comments to the existing code, and changes the exception
thrown when the user ACE exists into an information debug message.

Signed-off-by: Russ Whitehead <russwhitehead@yahoo.com>