HPCC-28617 Generate useFtSlave environment option in dfuserver config
The useFtSlave option which toggles whether ftslave or
dafilesrv is used as the spray service by dfuserver was
being ignored if it was in environment.xml and not being
generated into the dfuserver config.

Consequently it was not possible to switch to using dafilesrv
as the spray service in bare-metal.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>