HPCC-28656 Automate package build & upload on tag
The process is now automated using GitHub Actions, so that when a repository is tagged, a workflow is automatically triggered to build packages and upload them as assets to the corresponding GitHub Release. If the tag contains "RC", it will be marked as a pre-release. This improves the efficiency and reliability of the release process.

Signed-off-by: Michael Gardner <michael.gardner@lexisnexisrisk.com>
Signed-off-by: Gordon Smith <GordonJSmith@gmail.com>