HPCC-28454 Add IFile rename retry logic
Allow IFile::rename to retry on failure.
This is mainly to cope/diagnose an issue with failing renames
on Azure file systems.
By default the retry count will be 0, i.e. behave as before,
except in Thor where default will be 10.
Configurable per job, at a Thor level, or globally.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>