HPCC-28819 Fix use_method_name logic for esdl cli
Ensure that when use_method_name is set for an EsdlService that the wsdl sets
the wsdl:message/wsdl:part/@element name for each "SoapIn" message is set to
the name of the method. Generate one xsd:element corresponding to each method
using a given request structure, and set that element's name attribute to
the name of the method that uses it.

Signed-off-by: Terrence Asselin <terrence.asselin@lexisnexisrisk.com>