HPCC-28885 Fix CDfsLogicalFileName getScopes on planes
For an external ~plane file name, CDfsLogicalFileName was
failing to set 'tailpos' which causes getScopes() to return
no scopes.
NB: This was also true for other external file forms, i.e.
~remote:: and ~file::

Also, this PR ensures that the scopes after the external
file specifier and name are validated where they haven't
been so historically.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>