HPCC-28896 Fix regression for listing log files using FileList
Before the dropzone scope check is added to the ESP FileSpray.FileList,
the FileList can also be used to retrieve a list of log files. In this
PR, the function of retrieving a list of log files is added back to the
FileList.
1. Rename the addDropZoneFile() to addPhysicalFile(). Rename/modify the
getDropZoneFiles() to getPhysicalFiles(). Both functions can be used
for both dropzone files and log files.
2. If a user does not specify a dropzone name and the specified path is
a valid log path, the FileList calls the getPhysicalFiles() to retrieve
a list of log files.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>