HPCC-28995 Ensure accessed time set on publish
Remove individual use cases of setting access timestamp
when a logical file is created, and instead move to attach,
such that it is consistently updated on file publish.
Prior to this fix, all sprayed files, and some other cases,
were not having an access timestamp set, which meant in turn
that the file expiry feature could not know when to delete them
unless they'd be read by something else.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>