HPCC-29073 Ensure platform-build* rebuilds on BUILD_TAG change
Make sure that the git clone is redone if BUILD_TAG changes.
Prior to this change, the tag could change the clone would
complete, but could fail on checkout because the tag was not
present in the remote repo.
Pushing it to the remote repo. would not fix the issue, because
rebuilding the image would not re-perform the git clone because
the tag had not changed since that successful layer.
Change means it will not cache the successful step, unless the
git clone, checkout and submodule succeeed.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>