HPCC-29189 Check Dropzone scope access before spraying/despraying
Before spraying a file from a dropzone, we need to check whether the user
has the permission to access the dropzone file or not. Similar permission
check has to be done before despraying a file to a dropzone.

Revise based on review
1. Inside the CWsFileIOEx::validateDropZoneAccess(), replace the
getDLFNPermissions() call with the getDropZoneScopePermissions().
2. In the checkDZScopeAccessAndSetSpraySourceDFUFileSpec(), use
the dropzone name and the host which are read from the
readAndCheckSpraySourceReq().
3. Remove the srcXML.append('\0').
4. Move the rmfn declaration to closer to where 1st used.
5. Use StringArray::appendList to retrieve files from a csv string.
6. Change the code to support path/filex* format.

Revise based on review
1. Revise the validateDropZoneAccess().
2. Remove the srcip.
3. Roll back the changes related to setFromXML() and add
the comments about removing related code later.
4. Change the code for removing file name from path
before calling getDropZoneScopePermissions().

Revise based on review
1. Rename 2 getDropZoneScopePermissions().
2. In the validateDropZoneAccess(), call the getDZFileScopePermissions()
to check scope permission.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>