HPCC-29160 Fix oversized compressed write
The compressed writer can write files larger than necessary, due
to a bug in flush().

flush() ensures all content is committed to disk.
In the case of CCompressedFile it 1st ensures it's on a new block
boundary (1MB), before it flushes unwritten content to disk.
Prior to this change, it also did this when nothing
had been written since the last flush, meaning it would unnecessarily
ensure it was on the next block boundary by back filling the gap
behind it.

The net result on a typical write that called flush before
collecting stats, then again on close() was that it unnecessarily
rounded the last incomplete block (1MB) and committed to disk.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>