HPCC-29053 Add --init-publisher-wuid ecl command line option
Add --init-publisher-wuid option wherever the cli supports
--dfu-copy.  --init-publisher-wuid will preallocate the publisher
wuid and display it on the command line before running the rest of
the command.  This will allow the user to track the progress even
if the cli interface gets disconnected from the server before
the command completes.

This commit also improves the setting of the Publisher Workunit
job name making it easier to find and understand the context of
publisher workunits.

Signed-off-by: Anthony Fishbeck <anthony.fishbeck@lexisnexis.com>