HPCC-27526 Avoid reading any non leaf nodes when reading an index with no filter
Prep work refactoring that would be required if we were to try to make the loading
of the root node be on-demand. This change is a worthwhile code cleanup anyway, but
I'm not convinced it makes sense to complete the work since unfiltered index reading
is a very rare operation compared to filtered reads, which would incur an overhead
on every seek.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>