HPCC-27588 Add ESP generated form support for self referencing structs
This adds support for structures that reference themselves directly or
as an ArrayOf<> but doesn't support grandparent references. If needed we
can add support for that in the future.  One level of recursion was
challenging enough for the moment.

Also fixed the SOAP test page for the same type of structure.

Specifically targeting as a fix for ws_logaccess::GetLogs.

Signed-off-by: Anthony Fishbeck <anthony.fishbeck@lexisnexis.com>