HPCC-29365 Validate username for CRLF_Injection
When an ESP is running in the UserNameOnlyMode, the username
is stored inside a Cookie. The changes in this PR validate
the username. If an invalid username is found, the ESP will
ask for a username again.

Bad username may be sent in a url before the GetUserName page.

Revise based on review
1. Rename the isValidUserName() to isMalformedUserName().
2. Trim the username input.
3. Change the messages from 'invalid' to 'malformed'.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>