HPCC-29573 Fix sasha service not found in WS-FS
When I archive a DFU workunit in a cloud environment, I got the
'Service 'sasha-wu-archiver' not found' error. It is because
incorrect service type is used to find out sasha service. For ECL
WU, the service type for sasha should be 'wu-archiver'. For DFU
WU, the service type for sasha should be 'dfuwu-archiver'.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>