HPCC-29670 Fix broken file spray (no DZRestriction) and refactoring
This PR is a copy of the original branch for https://github.com/hpcc-systems/HPCC-Platform/pull/17378.
To minimize the changes for 9.0.x, the latest branch for that PR only
contains the fix for the crash when file spray with DropZoneRestriction
disabled. This PR replaces that PR for master.

1. When spraying a file with no DropZoneRestriction, the
getAndValidateDropZone() may not return a dropzone plane. The
readAndCheckSpraySourceReq() should validate the return and read the
plane name if there is the return.
2. In the checkDZScopeAccessAndSetSpraySourceDFUFileSpec(), validate
DZPathScopePermissions only if the dropzone plane name is found in
the readAndCheckSpraySourceReq().
3. Create throwOrLogDropZoneLookUpError(): if not bare metal or
DropZoneRestriction enabled, throw an exception; otherwise, log the
error.
4. Create findDropZonePlaneName(): if a plane name is not found based
on dropzone host and path, call throwOrLogDropZoneLookUpError().
5. In ESP FileSpray service, call the findDropZonePlaneName() to find
a dropzone plane name. Validate DZPathScopePermissions only if the
plane name is found.
6. In getDZPathScopePermissions(), if a dropzone plane is not found
based on dropzone host and path, call throwOrLogDropZoneLookUpError().

Revise based on review:
1. Remove the host from the getDZPathScopePermissions() and the
getDZFileScopePermissions().
2. Move the findDropZonePlaneName() so that it can be called from
both ws_fsService.cpp and ws_fsBinding.cpp.
3. Add checkDZPathScopePermissions() and checkDZFileScopePermissions().

Revise based on review:
1. Add isPathInPlane().
2. Remove the validateDropZonePath() and change related code.
3. Move the code from the throwOrLogDropZoneLookUpError() to the findDropZonePlaneName().
4. Several small changes.

Revise based on review:
1. Change the isPathInPlane() to return true for empty path.
2. Move the checkDropZoneInputAndAccess() to SMCLib.
3. Revise it so that it can be used by ws_fsService.cpp.
4. Revise the onFileList() to use checkDropZoneInputAndAccess().
5. Call checkDropZoneInputAndAccess() from OnFileList and
onDeleteDropZoneFiles.
6. Call checkDropZoneInput() from OnDespray.

Revise based on review:
1. Do not check 'localhost' if containerized
2. Revise readAndCheckSpraySourceReq() for spray to use
checkDropZoneInput(), including: move the scope access check from
checkDZScopeAccessAndSetSpraySourceDFUFileSpec()
to readAndCheckSpraySourceReq() and rename
checkDZScopeAccessAndSetSpraySourceDFUFileSpec()
to setSpraySourceDFUFileSpec().
3. Remove the #include "dautils.hpp" from ws_Binding.cpp since the
checkDropZoneInputAndAccess() has been moved to SMCLib.
4. Clean the code for the XML MemoryBuffer cast.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>